# api/src/slices/core/decorators/ApiListResponse.ts
import { Type, applyDecorators } from '@nestjs/common';
import { ApiExtraModels, ApiOkResponse, getSchemaPath } from '@nestjs/swagger';

export class ListModel<T> {
  public readonly data: T[];
  public readonly success: boolean;
}

export const ApiListResponse = <TModel extends Type<any>>(model: TModel) => {
  return applyDecorators(
    ApiExtraModels(ListModel, model),
    ApiOkResponse({
      schema: {
        allOf: [
          { $ref: getSchemaPath(ListModel) },
          {
            properties: {
              data: {
                type: 'array',
                items: { $ref: getSchemaPath(model) },
              },
            },
          },
        ],
      },
    }),
  );
};