# api/src/slices/users/apiKeys/dtos/create-api-key.dto.ts
import { IsString } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';
import { ICreateApiKeyData } from '../domain';
export class CreateApiKeyDto implements ICreateApiKeyData {
  teamId: string;

  @ApiProperty()
  @IsString()
  name: string;
}